backend/requirements.txt
asgiref==3.10.0
Django==4.2.25
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0

django-filter==24.3

psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
